Move sparsity layers to be part of quantization layers; to enable applying both the techniques together for finetuning/one-shot application.
Currently sparsity is supported for Linear and Attention layers only, but can be easily extended to other layers.

This does not change the existing behavior of quantization layers and models using quantized layers. Sparsity functions are no-op by default, and would sparsify weights by setting argument sparsity.

Will add documentation on how to apply sparsity. Changes sparsity_helper and other usage of sparsity layers accordingly to enable sparsity for a model using quantization layers.

PiperOrigin-RevId: 564506012
